+++ common/src/commonMain/kotlin/com/powersync/db/PowerSyncDatabaseImpl.kt
@@ -1,6 +1,7 @@
 package com.powersync.db
 
 import co.touchlab.kermit.Logger
+import com.powersync.ExperimentalCheckpointRequestsApi
 import com.powersync.ExperimentalPowerSyncAPI
 import com.powersync.PowerSyncDatabase
 import com.powersync.PowerSyncException
@@ -8,6 +9,7 @@ import com.powersync.bucket.BucketStorage
 import com.powersync.bucket.BucketStorageImpl
 import com.powersync.bucket.StreamPriority
 import com.powersync.bucket.targetCheckpointRequestId
+import com.powersync.connectors.CustomCheckpointRequestConnector
 import com.powersync.connectors.PowerSyncBackendConnector
 import com.powersync.db.crud.CrudBatch
 import com.powersync.db.crud.CrudEntry
@@ -19,6 +21,7 @@ import com.powersync.db.internal.InternalDatabaseImpl
 import com.powersync.db.internal.InternalTable
 import com.powersync.db.internal.PowerSyncVersion
 import com.powersync.db.schema.Schema
+import com.powersync.sync.CheckpointMode
 import com.powersync.sync.CoreSyncStatus
 import com.powersync.sync.StreamingSyncClient
 import com.powersync.sync.SyncOptions
@@ -156,6 +159,13 @@ internal class PowerSyncDatabaseImpl(
             disconnectInternal()
 
             connectInternal { scope ->
+                @OptIn(ExperimentalCheckpointRequestsApi::class)
+                if (connector is CustomCheckpointRequestConnector && options.checkpointMode == CheckpointMode.Legacy) {
+                    logger.w {
+                        "A CustomCheckpointRequestConnector was used with legacy checkpoints, postCheckpointRequest will not get called"
+                    }
+                }
+
                 StreamingSyncClient(
                     status = currentStatus,
                     bucketStorage = bucketStorage,